Pull ShaderComponent and DataSourceComponent out to files.
